requirements.txt
ollama==0.4.4
requests==2.32.3
pydantic>=2.12.0
rich==13.9.4
python-dotenv==1.0.1
